services/llm_service/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "rokulkarni15", email = "rokulkarni015@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.36.0",
]

[project.scripts]
llm-service = "llm_service:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
